pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.84.0",
    "chromadb>=1.5.4",
    "langgraph>=0.6.11",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.50.0",
    "yfinance>=1.2.0",
]
